+++ cmem_plugin_kafka/constants.py
@@ -98,7 +98,7 @@ If 0 or less, all messages will be fetched.
 """
 
 DISABLE_COMMIT_DESCRIPTION = """
-Setting this to true will disable committing messages after retrival.
+Setting this to true will disable committing messages after retrieval.
 
 This means you will receive the same messages on the next execution (for debugging).
 """
@@ -197,9 +197,10 @@ PARSE_XML_LINK = "XmlParserOperator.md"
 
 PLUGIN_DOCUMENTATION = f"""
 This workflow operator uses the Kafka Consumer API
-to receive messages from an [Apache Kafka](https://kafka.apache.org/) topic.
+to receive messages from an [Apache Kafka](https://kafka.apache.org/) topic. It has
+no input port: it starts a workflow by reading from the topic on its own.
 
-Messages received from the topic will be generated as entities with the following
+By default, each message received is generated as one entity with the following
 flat schema:
 
 - **key** - the optional key of the message,
@@ -214,7 +215,8 @@ In order to process the resulting entities, they have to run through a transform
 
 As an alternate working mode, messages can be exported directly to a JSON or XML
 dataset if you know that the messages on your topic are valid JSON or XML documents
-(see Advanced Options > Messages Dataset).
+(see the Messages Dataset parameter). In this mode the task has no output port either
+- messages are written directly to the dataset instead of being handed on as entities.
 
 In this case, a sample response from the consumer will appear as follows: